// src/models/CreateItemParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateItemParams = {
    categoryId: number;
    genderId?: number;
    name: string;
    description?: string;
    image: string;
    frontImage?: string;
    backImage?: string;
    maskImage?: string;
    lootboxId: number;
};